requirements.txt
yfinance
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
plotly
statsmodels
cvxopt
cvxpy
requests
tqdm
streamlit
dash
dash-core-components
dash-html-components
pytest
python-dotenv
PyYAML